requirements.txt
flask
openai-whisper
requests
torch
